move clustermanager
